1、抽离登录组件，目前弹框登录正常
